Simplify the narrow implementation.
